Move arguments for ffmpeg from TransferFFmpegCore.cs in VideoModel.cs. Delete Obsolete message and clear code.

+++ Transfer.Game/IO/AudioExtract.cs
@@ -15,10 +15,21 @@ using Transfer.Game.Extensions;
 
 namespace Transfer.Game.IO;
 
-public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAudioExtract<T>
+public class AudioExtract<T> : IAudioExtract<T>
     where T : Track
 {
-    private readonly TransferConfigManager transferConfigManager1 = transferConfigManager ?? throw new ArgumentNullException(nameof(transferConfigManager));
+    private readonly TransferConfigManager transferConfigManager1;
+
+    private static readonly string tempFolder = Path.Combine(Path.GetTempPath(), "Transfer");
+    private readonly Storage tempStorage;
+
+    public AudioExtract(TransferConfigManager transferConfigManager)
+    {
+        transferConfigManager1 = transferConfigManager ?? throw new ArgumentNullException(nameof(transferConfigManager));
+        if (!Directory.Exists(tempFolder))
+            Directory.CreateDirectory(tempFolder);
+        tempStorage = new NativeStorage(tempFolder);
+    }
 
     public async Task<T> CreateaAndGetTrackAsync(string path, Storage storage, AudioManager audioManager, string outputPath = null, string arguments = null)
     {
@@ -33,9 +44,15 @@ public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAud
             return getTrackFromStorage(audioManager, resourceStore, outputPath);
         }
 
+        if (File.Exists(Path.Combine(tempFolder, outputPath)))
+        {
+            return getTrackFromStorage(audioManager, new StorageBackedResourceStore(tempStorage), outputPath);
+        }
+
         try
         {
-            string pathToFile = await convertFileAsync(path, "mp3", storage.GetFullPath(""));
+            string pathToFile = await convertFileAsync(path, "mp3", storage.GetFullPath("")) + ".mp3";
+            Logger.Log($"Path to file: {pathToFile}\n output path: {outputPath}");
             await saveFileToStorageAsync(pathToFile, storage, outputPath);
             File.Delete(pathToFile);
 
@@ -44,7 +61,7 @@ public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAud
                 throw new FileNotFoundException("Audio not found");
             }
 
-            return getTrackFromStorage(audioManager, resourceStore, outputPath);
+            return await CreateaAndGetTrackAsync(path, storage, audioManager, outputPath);
         }
         catch (FFMpegException)
         {
@@ -105,7 +122,7 @@ public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAud
 
     private async Task<string> convertFileAsync(string path, string extension, string outputPath = null, string arguments = null)
     {
-        Logger.Log($"Path is null: {path is null}({path})\n Extension is null: {extension is null}\n Output is null: {outputPath is null}\n Arguments: {arguments is null}");
+        Logger.Log($"Path is null: {path is null}({path})\n Extension is null: {extension is null}\n Output is null: {outputPath is null}({outputPath})\n Arguments: {arguments is null}");
         outputPath ??= Path.Combine(Path.GetTempPath(), $"{Guid.NewGuid()}.{extension}");
 
         if (!File.Exists(path))
@@ -119,9 +136,11 @@ public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAud
         return pathToFile;
     }
 
-    private async Task saveFileToStorageAsync(string pathToFile, Storage storage, string name)
+    private async Task saveFileToStorageAsync(string paToFile, Storage storage, string name)
     {
-        using (Stream file = new FileStream(pathToFile, FileMode.Open))
+        Logger.Log($"Path to file is null: {paToFile}\n Name is null: {name}");
+
+        using (Stream file = new FileStream(paToFile, FileMode.Open))
         {
             await saveStreamToStorageAsync(file, storage, name);
         }
@@ -129,10 +148,23 @@ public class AudioExtract<T>(TransferConfigManager transferConfigManager) : IAud
 
     private async Task saveStreamToStorageAsync(Stream file, Storage storage, string fileName)
     {
-        using (var storageStream = storage.GetStream(fileName, FileAccess.Write, FileMode.OpenOrCreate))
+        try
         {
-            Logger.Log($"File saved to {storage.GetFullPath(fileName)}");
-            await file.CopyToAsync(storageStream);
+            using (var storageStream = storage.GetStream(fileName, FileAccess.Write, FileMode.OpenOrCreate))
+            {
+                Logger.Log($"File saved to {storage.GetFullPath(fileName)}");
+                await file.CopyToAsync(storageStream);
+            }
+        }
+        catch (UnauthorizedAccessException)
+        {
+            Logger.Log("Access denied. Connecting to another Storage");
+
+            using (var storageStream = tempStorage.GetStream(fileName, FileAccess.Write, FileMode.OpenOrCreate))
+            {
+                Logger.Log($"File saved to {storage.GetFullPath(fileName)}");
+                await file.CopyToAsync(storageStream);
+            }
         }
     }
